requirements.txt
crewai==0.11.0
together==1.2.0
python-dotenv==1.0.1
pydantic==2.8.2
flask==3.0.0
gunicorn==21.2.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
